Core: TryRealloc improved
git-svn-id: svn://ultimatepp.org/upp/trunk@8099 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/sheap.cpp
@@ -247,7 +247,7 @@ bool Heap::TryRealloc(void *ptr, size_t newsize)
 	if((((dword)(uintptr_t)ptr) & 8) == 0) {
 		Page *page = (Page *)((uintptr_t)ptr & ~(uintptr_t)4095);
 		int k = page->klass;
-		return newsize <= Ksz(k);
+		return newsize <= (size_t)Ksz(k);
 	}
 	return LTryRealloc(ptr, newsize);
 }
